Additional DXGI stuff

+++ src/Titan.Windows/DXGI/LUID.cs
@@ -5,6 +5,6 @@ namespace Titan.Windows.DXGI;
 [StructLayout(LayoutKind.Sequential)]
 public struct LUID
 {
-    public ulong LowPart;
-    public long HighPart;
+    public uint LowPart;
+    public int HighPart;
 }
